pyproject.toml
[tool.poetry]
name = "interview-ai"
version = "0.1.0"
description = ""
authors = ["Daniel Lee"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
openai = "^1.3.5"
streamlit = "^1.27.0"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-multipart = "^0.0.6"
websockets = "^11.0.3"
httpx = "^0.25.0"
asyncio = "^3.4.3"
ffmpeg = "^1.4"
respx = "^0.20.2"
supabase = "^2.3.1"
argon2-cffi = "^23.1.0"
cryptography = "^41.0.7"
psycopg = {extras = ["binary"], version = "^3.1.17"}
redis = "^5.0.1"
pyjwt = "^2.8.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
